Auditor: finish refactor 1/x
