+++ auditor_refactored/vulnerability_catcher.py
@@ -1,8 +1,8 @@
-import security_settings
+from security_settings import secrets_pattern, dangerous_events
 import re
 
 def is_trigger_dangerous(trigger_name: str) -> bool:
-    return trigger_name in security_settings.dangerous_events
+    return trigger_name in dangerous_events
 
 
 def get_dangerous_triggers(triggers: list) -> list:
@@ -20,7 +20,7 @@ def list_secrets_used(full_yaml: str) -> list:
     :param full_yaml: The full yaml string of the workflow.
     :return: A list of all secrets being used in this workflow."""
     found_matches = []
-    secrets = re.compile(security_settings.secrets_pattern)
+    secrets = re.compile(secrets_pattern)
     if matches:= secrets.findall(full_yaml):
         for match in matches:
             if match not in found_matches:
